pyproject.toml
[project]
name = "vibecheck"
version = "0.1.0"
description = "Agentic prototype for vibe-based venue recommendations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.186.1",
    "crewai-tools>=0.71.0",
    "gradio>=5.45.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "jupyter>=1.1.1",
]
